TF/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
faiss-cpu>=1.7.4
redis>=5.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

